@@ -12,7 +12,7 @@
         }
     ],
     "require": {
-        "php": ">=7.2.0",
+        "php": ">=7.4.0",
         "ParticleBits/pdo": "~2.0"
     },
     "autoload": {
